PLUGINAPI-182 Create a shared module for the Version class
