--- manifest/src/test/java/org/sonar/plugin/api/manifest/VersionTests.java
@@ -1,262 +1,0 @@
-/*
- * Sonar Plugin API
- * Copyright (C) 2009-2025 SonarSource Sàrl
- * mailto:info AT sonarsource DOT com
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 3 of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
- */
-package org.sonar.plugin.api.manifest;
-
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-class VersionTests {
-
-  @Test
-  void test_fields_of_snapshot_versions() {
-    var version = Version.create("1.2.3-SNAPSHOT");
-    assertThat(version.getMajor()).isEqualTo(1);
-    assertThat(version.getMinor()).isEqualTo(2);
-    assertThat(version.getPatch()).isEqualTo(3);
-    assertThat(version.getBuild()).isZero();
-    assertThat(version.getQualifier()).isEqualTo("SNAPSHOT");
-  }
-
-  @Test
-  void test_fields_of_releases() {
-    var version = Version.create("1.2");
-    assertThat(version.getMajor()).isEqualTo(1);
-    assertThat(version.getMinor()).isEqualTo(2);
-    assertThat(version.getPatch()).isZero();
-    assertThat(version.getBuild()).isZero();
-    assertThat(version.getQualifier()).isEmpty();
-  }
-
-  @Test
-  void compare_releases() {
-    var version12 = Version.create("1.2");
-    var version121 = Version.create("1.2.1");
-
-    assertThat(version12)
-      .hasToString("1.2")
-      .isEqualByComparingTo(version12);
-    assertThat(version121)
-      .isEqualByComparingTo(version121)
-      .isGreaterThan(version12);
-  }
-
-  @Test
-  void compare_snapshots() {
-    var version12 = Version.create("1.2");
-    var version12Snapshot = Version.create("1.2-SNAPSHOT");
-    var version121Snapshot = Version.create("1.2.1-SNAPSHOT");
-    var version12RC = Version.create("1.2-RC1");
-
-    assertThat(version12).isGreaterThan(version12Snapshot);
-    assertThat(version12Snapshot).isEqualByComparingTo(version12Snapshot);
-    assertThat(version121Snapshot).isGreaterThan(version12Snapshot);
-    assertThat(version12Snapshot).isGreaterThan(version12RC);
-  }
-
-  @Test
-  void compare_release_candidates() {
-    var version12 = Version.create("1.2");
-    var version12Snapshot = Version.create("1.2-SNAPSHOT");
-    var version12RC1 = Version.create("1.2-RC1");
-    var version12RC2 = Version.create("1.2-RC2");
-
-    assertThat(version12RC1)
-      .isLessThan(version12Snapshot)
-      .isEqualByComparingTo(version12RC1)
-      .isLessThan(version12RC2)
-      .isLessThan(version12);
-  }
-
-  @Test
-  void testTrim() {
-    var version12 = Version.create("   1.2  ");
-
-    assertThat(version12.getName()).isEqualTo("1.2");
-    assertThat(version12).isEqualTo(Version.create("1.2"));
-  }
-
-  @Test
-  void testDefaultNumberIsZero() {
-    var version12 = Version.create("1.2");
-    var version120 = Version.create("1.2.0");
-
-    assertThat(version12).isEqualTo(version120);
-    assertThat(version120).isEqualTo(version12);
-  }
-
-  @Test
-  void testCompareOnTwoDigits() {
-    var version1dot10 = Version.create("1.10");
-    var version1dot1 = Version.create("1.1");
-    var version1dot9 = Version.create("1.9");
-
-    assertThat(version1dot10)
-      .isGreaterThan(version1dot1)
-      .isGreaterThan(version1dot9);
-  }
-
-  @Test
-  void testFields() {
-    var version = Version.create("1.10.2");
-
-    assertThat(version.getName()).isEqualTo("1.10.2");
-    assertThat(version).hasToString("1.10.2");
-    assertThat(version.getMajor()).isEqualTo(1);
-    assertThat(version.getMinor()).isEqualTo(10);
-    assertThat(version.getPatch()).isEqualTo(2);
-    assertThat(version.getBuild()).isZero();
-  }
-
-  @Test
-  void testPatchFieldsEquals() {
-    var version = Version.create("1.2.3.4");
-
-    assertThat(version.getPatch()).isEqualTo(3);
-    assertThat(version.getBuild()).isEqualTo(4);
-
-    assertThat(version)
-      .isEqualTo(version)
-      .isEqualTo(Version.create("1.2.3.4"))
-      .isNotEqualTo(Version.create("1.2.3.5"));
-  }
-
-  @Test
-  void removeQualifier() {
-    var version = Version.create("1.2.3-SNAPSHOT").removeQualifier();
-
-    assertThat(version.getMajor()).isEqualTo(1);
-    assertThat(version.getMinor()).isEqualTo(2);
-    assertThat(version.getPatch()).isEqualTo(3);
-    assertThat(version.getQualifier()).isEmpty();
-  }
-
-  @Test
-  void removeQualifier_when_no_qualifier() {
-    var version = Version.create("1.2.3").removeQualifier();
-
-    assertThat(version.getMajor()).isEqualTo(1);
-    assertThat(version.getMinor()).isEqualTo(2);
-    assertThat(version.getPatch()).isEqualTo(3);
-    assertThat(version.getQualifier()).isEmpty();
-    assertThat(version.getName()).isEqualTo("1.2.3");
-  }
-
-  @Test
-  void satisfiesMinRequirement_should_return_true_when_equal() {
-    var version = Version.create("1.2.3");
-    var minRequirement = Version.create("1.2.3");
-
-    assertThat(version.satisfiesMinRequirement(minRequirement)).isTrue();
-  }
-
-  @Test
-  void satisfiesMinRequirement_should_return_true_when_greater() {
-    var version = Version.create("1.3.0");
-    var minRequirement = Version.create("1.2.3");
-
-    assertThat(version.satisfiesMinRequirement(minRequirement)).isTrue();
-  }
-
-  @Test
-  void satisfiesMinRequirement_should_return_false_when_lower() {
-    var version = Version.create("1.2.2");
-    var minRequirement = Version.create("1.2.3");
-
-    assertThat(version.satisfiesMinRequirement(minRequirement)).isFalse();
-  }
-
-  @Test
-  void satisfiesMinRequirement_should_ignore_qualifier() {
-    var version = Version.create("1.2.3-SNAPSHOT");
-    var minRequirement = Version.create("1.2.3-RC1");
-
-    assertThat(version.satisfiesMinRequirement(minRequirement)).isTrue();
-  }
-
-  @Test
-  void compareToIgnoreQualifier_should_return_zero_when_equal() {
-    var version1 = Version.create("1.2.3-SNAPSHOT");
-    var version2 = Version.create("1.2.3-RC1");
-
-    assertThat(version1.compareToIgnoreQualifier(version2)).isZero();
-  }
-
-  @Test
-  void compareToIgnoreQualifier_should_return_positive_when_greater() {
-    var version1 = Version.create("1.3.0-SNAPSHOT");
-    var version2 = Version.create("1.2.3-SNAPSHOT");
-
-    assertThat(version1.compareToIgnoreQualifier(version2)).isPositive();
-  }
-
-  @Test
-  void compareToIgnoreQualifier_should_return_negative_when_lower() {
-    var version1 = Version.create("1.2.2-SNAPSHOT");
-    var version2 = Version.create("1.2.3-SNAPSHOT");
-
-    assertThat(version1.compareToIgnoreQualifier(version2)).isNegative();
-  }
-
-  @Test
-  void equals_should_return_false_for_null() {
-    var version = Version.create("1.2.3");
-
-    assertThat(version.equals(null)).isFalse();
-  }
-
-  @Test
-  void equals_should_return_false_for_different_type() {
-    var version = Version.create("1.2.3");
-
-    assertThat(version.equals("1.2.3")).isFalse();
-  }
-
-  @Test
-  void equals_should_return_true_for_same_instance() {
-    var version = Version.create("1.2.3");
-
-    assertThat(version.equals(version)).isTrue();
-  }
-
-  @Test
-  void hashCode_should_be_consistent() {
-    var version1 = Version.create("1.2.3-SNAPSHOT");
-    var version2 = Version.create("1.2.3-SNAPSHOT");
-
-    assertThat(version1).hasSameHashCodeAs(version2);
-  }
-
-  @Test
-  void hashCode_should_differ_for_different_versions() {
-    var version1 = Version.create("1.2.3");
-    var version2 = Version.create("1.2.4");
-
-    assertThat(version1.hashCode()).isNotEqualTo(version2.hashCode());
-  }
-
-  @Test
-  void hashCode_should_differ_for_different_qualifiers() {
-    var version1 = Version.create("1.2.3-SNAPSHOT");
-    var version2 = Version.create("1.2.3-RC1");
-
-    assertThat(version1.hashCode()).isNotEqualTo(version2.hashCode());
-  }
-}
